Revert "Fix : LT-18440 Convert Kernel ProjectCode Support 64bit"
This reverts commit ce4a4615364eacad612e807b8210e885237cd637.

+++ Src/Generic/HashMap.cpp
@@ -213,7 +213,7 @@ int HashChars::operator () (const char * pszKey)
 {
 	if (NULL == pszKey)
 		return 0;
-	int cb = (int)(strlen(pszKey));
+	int cb = strlen(pszKey);
 	if (0 == cb)
 		return 0;
 	int nHash = 0;
